requirements.txt
streamlit
requests
torch
torchvision
ftfy
regex
tqdm
Pillow
clip